python3-botocore: use nobranch=1 to reflect the latest repo status
Issue: LINCD-18987

The rev is a valid tag but not in any branch. This is the current
status of botocore repo. So use nobranch=1 to reflect such status.
This fixes do_fetch error.

Upstream PR: https://github.com/aws4embeddedlinux/meta-aws/pull/9393
(LOCAL REV: NOT UPSTREAM) -- sent to upstream on 20240821

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>